homework2/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
behave==1.3.3
behave-django==1.9.0
betterpath==0.2.2
click==8.3.1
colorama==0.4.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
Django==6.0.2
djangorestframework==3.16.1
l==0.11.0
parse==1.21.1
parse_type==0.6.6
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
vcversioner==2.16.0.0
zope.interface==8.2
